Just made Paprika Better.
So, Paprika now decrypts to a folder in the current directory and
does not delete the encrypted images --- thus saving them in case
of a wrong pass-file use.
As soon as I'd fixed that there was a bug. The bug made it so that
if there was a directory in the current dir with 'image' in it's
name, it would cause the dir to be passed to convert and convert
would complain. Fixed that issue by inserting a check that checks
whether or not the file in question is a directory.
I think there's still a bug. If I name a non-image file to be 'im-
age', the bug just might reappear. Need a fix for that.